fix(aria-input-field-name): skip combobox popups
